Fix train/inference masking mismatch, add self-revision sampling, clean sanity-check dataset
- train.py: remove apply_partial_reveal, which always revealed 50% of masked
  positions back to ground truth regardless of t. That made the model train
  on less corruption than sample.py actually shows it at generation time for
  the same t, worst at high-t (heavily masked) steps. Loss is now computed
  directly on diffusion.corrupt's own is_masked, matching inference exactly.
- diffusion.py: fix TokenDifficulty.update crashing on GPU -- it ran
  cross-entropy on raw fp16 AMP logits then scatter_add'ed into an fp32
  accumulator, a dtype mismatch never hit before because this path had never
  actually been run end-to-end on CUDA.
- sample.py: sampler can now revise its own earlier guesses. Each step,
  already-revealed positions the model is no longer confident in get
  remasked and reconsidered on a later step with fresher context, instead of
  being frozen the instant they're first revealed. This is the structural
  difference from an autoregressive model, which can never revisit a token.
  Remasking is capped, decays by the final step, and is excluded from that
  same step's reveal pool -- convergence to a fully unmasked sequence is
  still guaranteed. New --remask-threshold/--max-remask-frac flags in
  generate.py.
- sanity_check.py: switch default dataset from wikitext-2-raw-v1 to
  roneneldan/TinyStories -- clean plain prose instead of wiki markup
  artifacts, better signal for a quick "does the pipeline learn" check.
- Verified locally end-to-end: 10 epochs on TinyStories, val loss
  5.82 -> 4.35, decoded samples are grammatical short stories.
- README.md / AdaMask_Colab_Quickstart.ipynb updated to match.

+++ adamask/diffusion.py
@@ -61,7 +61,10 @@ class TokenDifficulty(MaskedDiffusion):
             return
 
         # Per-token cross-entropy: higher loss means the token is harder.
-        per_token_loss = F.cross_entropy(masked_logits, target, reduction="none")
+        # logits arrive in whatever dtype the forward pass ran in (fp16 under
+        # AMP), but self.total/self.loss_sum are fp32 accumulators -- cast
+        # here so scatter_add_ below doesn't hit a dtype mismatch.
+        per_token_loss = F.cross_entropy(masked_logits.float(), target, reduction="none")
 
         # Decay existing stats, then accumulate this batch's counts/loss on top.
         # Dividing two accumulators decayed by the same factor each step is
